+++ configuration/configuration_test.go
@@ -14,10 +14,10 @@ package configuration_test
 import (
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/tideland/golib/audit"
 	"github.com/tideland/golib/configuration"
+	"github.com/tideland/golib/stringex"
 )
 
 //--------------------
@@ -79,8 +79,8 @@ func TestList(t *testing.T) {
 	assert.ErrorMatch(err, `.* invalid configuration path "/config/x"`)
 }
 
-// TestGetSuccess tests the successful retrieval of values.
-func TestGetSuccess(t *testing.T) {
+// TestValueSuccess tests the successful retrieval of values.
+func TestValueSuccess(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
 
 	source := `{config
@@ -88,8 +88,6 @@ func TestGetSuccess(t *testing.T) {
 	{b true}
 	{c -1}
 	{d     47.11     }
-	{e 2015-06-25T23:45:00+02:00}
-	{f 2h15m30s}
 	{sub
 		{a
 			World}
@@ -98,33 +96,30 @@ func TestGetSuccess(t *testing.T) {
 	config, err := configuration.Read(strings.NewReader(source))
 	assert.Nil(err)
 
-	vs, err := config.At("a").Get()
+	v, err := config.At("a").Value()
 	assert.Nil(err)
-	assert.Equal(vs, "Hello")
-	vb, err := config.At("b").GetBool()
+	assert.Equal(v, "Hello")
+	v, err = config.At("b").Value()
 	assert.Nil(err)
-	assert.Equal(vb, true)
-	vi, err := config.At("c").GetInt()
+	assert.Equal(v, "true")
+	v, err = config.At("c").Value()
 	assert.Nil(err)
-	assert.Equal(vi, -1)
-	vd, err := config.At("d").GetFloat64()
+	assert.Equal(v, "-1")
+	v, err = config.At("d").Value()
 	assert.Nil(err)
-	assert.Equal(vd, 47.11)
-	vtim, err := config.At("e").GetTime()
+	assert.Equal(v, "47.11")
+	v, err = config.At("sub", "a").Value()
 	assert.Nil(err)
-	loc, err := time.LoadLocation("CET")
+	assert.Equal(v, "World")
+	v, err = config.At("sub", "b").Value()
 	assert.Nil(err)
-	assert.Equal(vtim.String(), time.Date(2015, time.June, 25, 23, 45, 00, 0, loc).String())
-	vdur, err := config.At("f").GetDuration()
-	assert.Nil(err)
-	assert.Equal(vdur, 8130*time.Second)
+	assert.Equal(v, "42")
 
-	vs, err = config.At("sub", "a").Get()
-	assert.Nil(err)
-	assert.Equal(vs, "World")
-	vi, err = config.At("sub", "b").GetInt()
-	assert.Nil(err)
-	assert.Equal(vi, 42)
+	d := stringex.NewDefaulter("config", true)
+	vi := d.AsInt(config.At("c"), 42)
+	assert.Equal(vi, -1)
+	vf := d.AsFloat64(config.At("d"), 12.34)
+	assert.Equal(vf, 47.11)
 }
 
 // TestGetFail tests the failing retrieval of values.
@@ -135,47 +130,10 @@ func TestGetFail(t *testing.T) {
 	config, err := configuration.Read(strings.NewReader(source))
 	assert.Nil(err)
 
-	_, err = config.At("x").Get()
+	_, err = config.At("x").Value()
 	assert.ErrorMatch(err, `.* invalid configuration path "/config/x"`)
-	_, err = config.At("sub", "x").Get()
+	_, err = config.At("sub", "x").Value()
 	assert.ErrorMatch(err, `.* invalid configuration path "/config/sub/x"`)
-
-	vb, err := config.At("a").GetBool()
-	assert.Equal(vb, false)
-	assert.ErrorMatch(err, `.* invalid value format of "Hello": strconv.ParseBool: .*`)
-	vi, err := config.At("a").GetInt()
-	assert.Equal(vi, 0)
-	assert.ErrorMatch(err, `.* invalid value format of "Hello": strconv.ParseInt: .*`)
-	vf, err := config.At("a").GetFloat64()
-	assert.Equal(vf, 0.0)
-	assert.ErrorMatch(err, `.* invalid value format of "Hello": strconv.ParseFloat: .*`)
-}
-
-// TestGetDefault tests the retrieval of values with defaults.
-func TestGetDefault(t *testing.T) {
-	assert := audit.NewTestingAssertion(t, true)
-
-	source := "{config {a Hello}}"
-	config, err := configuration.ReadString(source)
-	assert.Nil(err)
-
-	vs := config.At("a").GetDefault("default")
-	assert.Equal(vs, "Hello")
-
-	vs = config.At("foo").GetDefault("default")
-	assert.Equal(vs, "default")
-	vb := config.At("foo").GetBoolDefault(true)
-	assert.True(vb)
-	vi := config.At("foo").GetIntDefault(42)
-	assert.Equal(vi, 42)
-	vf := config.At("foo").GetFloat64Default(47.11)
-	assert.Equal(vf, 47.11)
-	now := time.Now()
-	vt := config.At("foo").GetTimeDefault(now)
-	assert.Equal(vt, now)
-	dur := 5 * time.Second
-	vd := config.At("foo").GetDurationDefault(dur)
-	assert.Equal(vd, dur)
 }
 
 // TestApply tests the applying of values.
@@ -191,12 +149,15 @@ func TestApply(t *testing.T) {
 		"b":     "42",
 	})
 	assert.Nil(err)
-	vs, err := applied.At("sub", "a").Get()
+	v, err := applied.At("sub", "a").Value()
 	assert.Nil(err)
-	assert.Equal(vs, "Tester")
-	vi, err := applied.At("b").GetInt()
+	assert.Equal(v, "Tester")
+	v, err = applied.At("a").Value()
 	assert.Nil(err)
-	assert.Equal(vi, 42)
+	assert.Equal(v, "Hello")
+	v, err = applied.At("b").Value()
+	assert.Nil(err)
+	assert.Equal(v, "42")
 
 }