+++ configuration/errors.go
@@ -31,7 +31,6 @@ var errorMessages = errors.Messages{
 	ErrIllegalSourceFormat: "illegal source format",
 	ErrIllegalConfigSource: "illegal source for configuration: %v",
 	ErrInvalidPath:         "invalid configuration path %q",
-	ErrInvalidFormat:       "invalid value format of %q",
 	ErrCannotApply:         "cannot apply values to configuration",
 }
 
@@ -42,12 +41,6 @@ var errorMessages = errors.Messages{
 // IsInvalidPathError checks if a path cannot be found.
 func IsInvalidPathError(err error) bool {
 	return errors.IsError(err, ErrInvalidPath)
-}
-
-// IsInvalidFormatError checks if a value hasn't the
-// expected format.
-func IsInvalidFormatError(err error) bool {
-	return errors.IsError(err, ErrInvalidFormat)
 }
 
 // EOF